Added Size Based Matching
Boxes now only attempt to match with others close to the same size,
which can be adjusted via slider. Added infrastructure for disparity to
depth calculations, but it currently doesn't yield good results since
rectified images have the same pixel locations. Also added a check of
the epipolar lines to see how good the remapping did... They should
remap to be perfectly horizontal (which the red lines are) if
rectification works correctly.