Added Overlap Box Constraint
Now limits the search region to only the bounding box of the area
visible to both cameras. Cuts down on bad matches, but still picks up
some incorrect ones that are outside the overlap but inside the bounding
box. Ideally, maximum interior rectangle would do better, but it doesn't
seem to be in opencv since orientation leads to a world of geometry
problems. However, there may be ways around this if we decided to push
this classifier further.