Harris Corner Detection
Added demo code to perform Harris corner detection. It's kinda glitchy,
sometimes spitting out an image totally filled with "identified
corners". This code is almost exactly the example code from the OpenCV
tutorial on Harris corner detection and I don't fully understand it at
this point.
The potentially really valuable addition is a function to find the
maximum interior rectangle for the overlap image. It makes the code have
a rather large startup, but will ensure that corners are only detected
within the region both cameras see and so no corners of undistorted
images get detected.